+++ src/main/java/com/sparta/bapzip/menu/application/MenuServiceV1.java
@@ -5,9 +5,9 @@ import com.sparta.bapzip.global.exception.GlobalException;
 import com.sparta.bapzip.menu.domain.entity.MenuEntity;
 import com.sparta.bapzip.menu.domain.enums.MenuStatus;
 import com.sparta.bapzip.menu.domain.repository.MenuRepository;
-import com.sparta.bapzip.menu.presentation.dto.request.MenuCreateRequest;
-import com.sparta.bapzip.menu.presentation.dto.request.MenuStatusUpdateRequest;
-import com.sparta.bapzip.menu.presentation.dto.request.MenuUpdateRequest;
+import com.sparta.bapzip.menu.application.dto.request.MenuCreateRequest;
+import com.sparta.bapzip.menu.application.dto.request.MenuStatusUpdateRequest;
+import com.sparta.bapzip.menu.application.dto.request.MenuUpdateRequest;
 import com.sparta.bapzip.menu.presentation.dto.response.MenuCreateResponse;
 import com.sparta.bapzip.menu.presentation.dto.response.MenuDetailResponse;
 import com.sparta.bapzip.menu.presentation.dto.response.MenuListByShopResponse;
