+++ src/main/java/com/sparta/bapzip/menu/application/dto/request/MenuStatusUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.sparta.bapzip.menu.presentation.dto.request;
+package com.sparta.bapzip.menu.application.dto.request;
 
 import jakarta.validation.constraints.NotNull;
