+++ src/main/java/com/sparta/bapzip/menu/application/dto/request/MenuUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.sparta.bapzip.menu.presentation.dto.request;
+package com.sparta.bapzip.menu.application.dto.request;
 
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
