+++ src/main/java/com/sparta/bapzip/menu/presentation/controller/MenuControllerV1.java
@@ -3,10 +3,10 @@ package com.sparta.bapzip.menu.presentation.controller;
 import com.sparta.bapzip.global.response.ApiResponse;
 import com.sparta.bapzip.global.response.PageResponseDto;
 import com.sparta.bapzip.menu.application.MenuServiceV1;
-import com.sparta.bapzip.menu.presentation.dto.request.MenuCreateRequest;
-import com.sparta.bapzip.menu.presentation.dto.request.MenuSearchRequest;
-import com.sparta.bapzip.menu.presentation.dto.request.MenuStatusUpdateRequest;
-import com.sparta.bapzip.menu.presentation.dto.request.MenuUpdateRequest;
+import com.sparta.bapzip.menu.application.dto.request.MenuCreateRequest;
+import com.sparta.bapzip.menu.application.dto.request.MenuSearchRequest;
+import com.sparta.bapzip.menu.application.dto.request.MenuStatusUpdateRequest;
+import com.sparta.bapzip.menu.application.dto.request.MenuUpdateRequest;
 import com.sparta.bapzip.menu.presentation.dto.response.MenuCreateResponse;
 import com.sparta.bapzip.menu.presentation.dto.response.MenuDetailResponse;
 import com.sparta.bapzip.menu.presentation.dto.response.MenuListByShopResponse;
@@ -14,7 +14,6 @@ import com.sparta.bapzip.menu.presentation.dto.response.MenuSearchResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
